+++ zsLib/cpp/zsLib_XMLParserPos.cpp
@@ -54,7 +54,7 @@ namespace zsLib
       }
 
       //-----------------------------------------------------------------------
-      static ULONG calculateColumnFromSOL(const char *const inPos, const char *const inStart, ULONG inTabSize)
+      static size_t calculateColumnFromSOL(const char *const inPos, const char *const inStart, ULONG inTabSize)
       {
         // find the start of the line
         const char *startLine = inPos;
@@ -276,7 +276,7 @@ namespace zsLib
     }
 
     //-------------------------------------------------------------------------
-    ParserPos &ParserPos::operator+=(ULONG inDistance)
+    ParserPos &ParserPos::operator+=(size_t inDistance)
     {
       ParserPtr parser(getParser());
 
@@ -337,7 +337,7 @@ namespace zsLib
     }
 
     //-------------------------------------------------------------------------
-    ParserPos &ParserPos::operator-=(ULONG inDistance)
+    ParserPos &ParserPos::operator-=(size_t inDistance)
     {
       ParserPtr parser(getParser());
 
@@ -345,7 +345,7 @@ namespace zsLib
       {
         ParserPos temp(*this);
         temp.setSOF();
-        temp += (ULONG)strlen(temp.mPos);
+        temp += strlen(temp.mPos);
         (*this) = temp;
       }
 
@@ -460,7 +460,7 @@ namespace zsLib
     }
 
     //-------------------------------------------------------------------------
-    ParserPos operator+(const ParserPos &inPos, ULONG inDistance)
+    ParserPos operator+(const ParserPos &inPos, size_t inDistance)
     {
       ParserPos temp(inPos);
       temp += inDistance;
@@ -468,7 +468,7 @@ namespace zsLib
     }
 
     //-------------------------------------------------------------------------
-    ParserPos operator-(const ParserPos &inPos, ULONG inDistance)
+    ParserPos operator-(const ParserPos &inPos, size_t inDistance)
     {
       ParserPos temp(inPos);
       temp -= inDistance;
@@ -478,25 +478,25 @@ namespace zsLib
     //-------------------------------------------------------------------------
     ParserPos operator+(const ParserPos &inPos, int inDistance)
     {
-      return inPos + ((ULONG)inDistance);
+      return inPos + ((size_t)inDistance);
     }
 
     //-------------------------------------------------------------------------
     ParserPos operator-(const ParserPos &inPos, int inDistance)
     {
-      return inPos - ((ULONG)inDistance);
+      return inPos - ((size_t)inDistance);
     }
 
     //-------------------------------------------------------------------------
     ParserPos operator+(const ParserPos &inPos, unsigned int inDistance)
     {
-      return inPos + ((ULONG)inDistance);
+      return inPos + ((size_t)inDistance);
     }
 
     //-------------------------------------------------------------------------
     ParserPos operator-(const ParserPos &inPos, unsigned int inDistance)
     {
-      return inPos - ((ULONG)inDistance);
+      return inPos - ((size_t)inDistance);
     }
   } // namespace XML
